beetle_backend/requirements.txt
numpy
sentence-transformers
qdrant-client
pydantic
bs4
langdetect
PyGithub
playwright
trafilatura
google-generativeai
dotenv